requirements.txt
numpy==1.22.4
pandas==1.1.3
Pillow==9.1.1
pycaret==2.3.10
streamlit==1.9.2



dash==2.3.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
gensim==3.8.3
geographiclib==1.52
geopy==2.2.0
gitdb==4.0.9
googletrans==4.0.0rc1
imbalanced-learn==0.7.0
importlib-metadata==4.8.2
importlib-resources==5.4.0
ipython-genutils==0.2.0
ipython-sql==0.4.0
ipywidgets == 7.5.1
jinja2==2.11.3
lightgbm==3.3.2
lyricsgenius==3.0.1
matplotlib==3.5.1
neo4j==4.4.3
neo4j-driver==4.3.6
neobolt==1.7.17
neomodel==4.0.8
nltk==3.5
numpy==1.22.4
pandas==1.1.3
Pillow==9.1.1
plotly==5.6.0
pmdarima==1.8.5
py2neo==2021.2.3
pycaret==2.3.10
seaborn==0.11.2
spacy==2.3.1
spacy-legacy==3.0.9
spacy-loggers==1.0.2
statsmodels==0.13.2
streamlit==1.9.2
textblob==0.17.1
wordcloud==1.8.1
xgboost==1.5.2
